+++ packages/core/application/application.android.ts
@@ -1,3 +1,4 @@
+import { embedded } from 'ui/embedding';
 import { profile } from '../profiling';
 import { View } from '../ui';
 import { AndroidActivityCallbacks, NavigationEntry } from '../ui/frame/frame-common';
@@ -9,6 +10,12 @@ declare namespace com {
 	namespace tns {
 		class NativeScriptApplication extends android.app.Application {
 			static getInstance(): NativeScriptApplication;
+		}
+
+		namespace embedding {
+			class ApplicationHolder {
+				static getInstance(): android.app.Application;
+			}
 		}
 	}
 }
@@ -356,6 +363,10 @@ export class AndroidApplication extends ApplicationCommon implements IAndroidApp
 		// check whether the com.tns.NativeScriptApplication type exists
 		if (com.tns.NativeScriptApplication) {
 			nativeApp = com.tns.NativeScriptApplication.getInstance();
+		}
+
+		if (!nativeApp && embedded()) {
+			nativeApp = com.tns.embedding.ApplicationHolder.getInstance();
 		}
 
 		// the getInstance might return null if com.tns.NativeScriptApplication exists but is not the starting app type