+++ packages/core/ui/gestures/touch-manager.ts
@@ -65,7 +65,7 @@ export class TouchManager {
 		const handleDown = (view?.touchAnimation && (<TouchAnimationOptions>view?.touchAnimation).down) || (TouchManager.animations && TouchManager.animations.down);
 		const handleUp = (view?.touchAnimation && (<TouchAnimationOptions>view?.touchAnimation).up) || (TouchManager.animations && TouchManager.animations.up);
 
-		if (global.isIOS) {
+		if (__IOS__) {
 			if (view?.ios?.addTargetActionForControlEvents) {
 				// can use UIControlEvents
 				if (!TouchManager.touchHandlers) {
@@ -215,7 +215,7 @@ export let TouchControlHandler: {
 ensureTouchControlHandlers();
 
 function ensureTouchControlHandlers() {
-	if (global.isIOS) {
+	if (__IOS__) {
 		@NativeClass
 		class TouchHandlerImpl extends NSObject {
 			private _owner: WeakRef<View>;