+++ packages/core/ui/text-base/index.d.ts
@@ -123,6 +123,11 @@ export class TextBase extends View implements AddChildFromBuilder {
 	/**
 	 * @private
 	 */
+	_getNativeTextTransform(value: CoreTypes.TextTransformType): android.text.method.TransformationMethod;
+
+	/**
+	 * @private
+	 */
 	_isSingleLine: boolean;
 	//@endprivate
 }